fix(lib): route on the full four-byte message code
Dispatch read one byte of the code, so a code with a non-zero high byte
was handled as its low byte. Message::get_message_code now reads the
whole little-endian uint32 and handlers register by u32. PeerInit and
PierceFirewall, the only frames with a one-byte code, get their own
accessor for the listener's first frame.

+++ soulseek-rs-lib/src/message/server/leave_room.rs
@@ -7,7 +7,7 @@ use std::sync::mpsc::Sender;
 pub struct LeaveRoomHandler;
 
 impl MessageHandler<ServerMessage> for LeaveRoomHandler {
-    fn get_code(&self) -> u8 {
+    fn get_code(&self) -> u32 {
         15
     }
